# ips_autobot_svr/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ips_autobot_svr API",
    "version": "1.0.0",
    "description": "Admin and user auth, logs, recharge APIs. Multi-app zhiling/yifei — pass app where required. Password fields use RSA encryption (GET /api/public-key)."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/api/admin/login": {
      "post": {
        "summary": "Admin login",
        "tags": [
          "Admin"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "passwordEncrypted"
                ],
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "passwordEncrypted": {
                    "type": "string",
                    "description": "RSA encrypted password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login success, returns JWT and admin info (includes admin_type)"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/api/admin": {
      "get": {
        "summary": "List admins (admin only)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "username",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Admin list (super admin only)"
          }
        }
      },
      "post": {
        "summary": "Create admin (admin only)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "passwordEncrypted"
                ],
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "admin_type": {
                    "type": "integer",
                    "enum": [
                      0,
                      1
                    ],
                    "description": "0=super admin, 1=normal admin"
                  },
                  "passwordEncrypted": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Admin created (super admin only)"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/admin/{id}": {
      "delete": {
        "summary": "Delete admin (admin only)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Admin deleted"
          },
          "400": {
            "description": "Cannot delete self or invalid id"
          },
          "404": {
            "description": "Admin not found"
          }
        }
      }
    },
    "/api/admin/password": {
      "put": {
        "summary": "Change current admin password",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "oldPasswordEncrypted",
                  "newPasswordEncrypted"
                ],
                "properties": {
                  "oldPasswordEncrypted": {
                    "type": "string"
                  },
                  "newPasswordEncrypted": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password updated"
          },
          "401": {
            "description": "Invalid old password"
          }
        }
      }
    },
...
    "/api/recharge": {
      "post": {
        "summary": "Create recharge record (admin or user)",
        "tags": [
          "Recharge"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "amount",
                  "app"
                ],
                "properties": {
                  "app": {
                    "type": "string",
                    "enum": [
                      "zhiling",
                      "yifei"
                    ],
                    "description": "Client app for recharge table"
                  },
                  "user_id": {
                    "type": "integer"
                  },
                  "username": {
                    "type": "string",
                    "description": "phone or username"
                  },
                  "amount": {
                    "type": "number"
                  },
                  "result": {
                    "type": "integer",
                    "enum": [
                      0,
                      1
                    ],
                    "default": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Recharge record created"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "get": {
        "summary": "List recharge records (admin sees all, user sees own)",
        "tags": [
          "Recharge"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "app",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "zhiling",
                "yifei"
              ]
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "user_id",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "result",
            "schema": {
              "type": "integer",
              "enum": [
                0,
                1
              ]
            }
          },
          {
            "in": "query",
            "name": "start_date",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "end_date",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated recharge records"
          }
        }
      }
    },
    "/api/score/change": {
      "post": {
        "summary": "Increase or deduct user score (admin only)",
        "tags": [
          "Score"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "change_type",
                  "tier_days",
                  "app"
                ],
                "properties": {
                  "app": {
                    "type": "string",
                    "enum": [
                      "zhiling",
                      "yifei"
                    ],
                    "description": "Client app for score columns and score_record"
                  },
                  "user_id": {
                    "type": "integer",
                    "description": "Target user id (or use phone)"
                  },
                  "phone": {
                    "type": "string",
                    "description": "Target user phone (or use user_id)"
                  },
                  "tier_days": {
                    "type": "integer",
                    "enum": [
                      30,
                      90,
                      180,
                      365
                    ],
                    "description": "Score tier days"
                  },
                  "score_change": {
                    "type": "integer",
                    "minimum": 1,
                    "description": "Optional override amount; defaults to selected tier score from users table for this app"
                  },
                  "change_type": {
                    "type": "integer",
                    "enum": [
                      0,
                      1
                    ],
                    "description": "0=add, 1=deduct"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Score changed and record created"
          },
          "400": {
            "description": "Bad request or insufficient score"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/api/score/records": {
      "get": {
        "summary": "List score records (admin sees all, user sees own)",
        "tags": [
          "Score"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "app",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "zhiling",
                "yifei"
              ]
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "user_id",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "username",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "phone",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "change_type",
            "schema": {
              "type": "integer",
              "enum": [
                0,
                1
              ]
            }
          },
          {
            "in": "query",
            "name": "start_date",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "end_date",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated score records"
          }
        }
      }
    },
...
  },
...
}